+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 ---------
 
 v2.0 (??? 2023)
+    * Use formatted strings where appropriate in std_io.
     * Removed exe_wrapper_v34 and hence support for < Python 3.7.
     * Use dict.get() function for keywords in wrappers.
     * Improved docstrings.